Keep the machine execution state in the state file
The machine execution state is one of ‘unknown’ (not checked or check
failed), ‘missing’ (i.e. VM not started yet or destroyed), ‘starting’
(VM creation, boot or reboot in progress), ‘up’ (i.e. reachable),
‘stopping’ (shutdown initiated) or ‘stopped’ (i.e. down but bootable).
‘charon info’ now displays these in conjunction with the deployment
state.  For instance, ‘Stopped / Up-to-date’ denotes that a machine is
currently stopped, but has the latest configuration.

The command ‘charon check’ is supposed to verify and update the
machine state in a backend-specific way.  For instance, if an EC2
machine has been stopped or terminated using the AWS command line
tools or console, then ‘charon check’ will discover this.

This is not entirely finished yet: the VirtualBox backend needs to
maintain the state, and the EC2 backend doesn't distinguish between
starting and up when doing ‘charon check’.