tests: Add basic VM test for Hetzner backend.
This test is currently more or less a no-op and currently resembles one
coordinator machine and two machines which (should) closely match the
Debian rescue systems used by Hetzner.

Unfortunately, work on this took longer than expected because I started
off with using live-boot from Debian Squeeze and ran through a big
minefield of bugs and glitches that were quite difficult to work around
and/or fix.

So, I went for using live-boot from a Wheezy image and creating a
Squeeze live ISO image from therein. Turns out that even that consumed
way too much time, because even there I stumbled on some bugs. The
latter are already fixed in Wheezy, but we can't use Wheezy for
simulating Hetzner's rescue systems, as those are still using Squeeze.

However, there are Wheezy rescue images available, but we can't boot
into them with the rescue API.

In the end, we now stick with Squeeze and we should now have a good
start for building our test sandbox. Fortunately, we're out of "hostile"
environments now and just need to monkey-patch the API access in order
to have our test environment set up.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>