nix_expr: Implement basic pretty printer.
This pretty printer now can handle everything but attrsets and it's
implemented in two phases: The first phase is producing all the syntax
we need and the second one only does indentation.

One major reason to do it like that is that we want to use multiline
strings whenever they're suitable (such as when the maximum line width
is exceeded) and avoid them if for example we just have a few \ns and
the list would fit inside the corresponding line.

Atom and Container are solely used for indentation and the reason
they're classes is because Python doesn't have ADT (or static typing at
all) and we want to build a nice tree of the indentation structure.

Basically, the structure of the tree is comprised of Container (which
acts more or less as a node, but with prefix/suffix and alternative
inline representation) and Atom, which is a single leaf of our tree.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>