Make sure that the machine is UP before trying to destroy it in GCP
This can cause problems since we issue API call without actually checking
if the node is up or not. One easy way to see the manifestation of this
bug is to create a deployment with a machine that has underscore in
it's  name (It's not supported in GCP). And without even deploying,
try to destroy it. The expected behavior is nothing! Since the
machine is down. However, an exception is raised because nixops tries
to call the API to get the node.