Remove the addressRange option from gce-networks and update API call
When trying to create a new network resource, the following error
message is returned by the API:

"Creation of legacy mode networks is deprecated. Please create a
subnet mode network instead by removing the IPv4Range field and adding
the autoCreateSubnetworks field to your network insert request."

This removes the `addressRange` option and updates the API call,
setting the `cidr` parameter to `None` and `mode` to `auto`.